+++ src/App/providers/AuthProvider/styles.module.css
@@ -12,3 +12,4 @@
         height: 6rem;
     }
 }
+
