+++ src/App/providers/CommandProvider/styles.module.css
@@ -1,0 +1,53 @@
+.last-saved-status {
+    display: flex;
+    position: fixed;
+    right: var(--spacing-md);
+    bottom: var(--spacing-md);
+    align-items: center;
+    transition: .5s opacity ease-in-out;
+    opacity: 0;
+    border-radius: 0.5rem;
+    background-color: var(--color-quaternary);
+    padding: var(--spacing-sm) var(--spacing-md);
+    color: var(--color-text-light);
+    gap: var(--spacing-xs);
+
+    .timur-icon {
+        height: 1rem;
+    }
+
+    &.active {
+        opacity: 1;
+
+        .timur-icon {
+            animation: shake-it-off 2.2s ease-in infinite;
+        }
+    }
+}
+
+@keyframes shake-it-off {
+    10% {
+        transform: rotate(0);
+    }
+    20% {
+        transform: rotate(-160deg);
+    }
+    30% {
+        transform: rotate(-160deg) translateY(2px);
+    }
+    35% {
+        transform: rotate(-160deg) translateY(-2px);
+    }
+    50% {
+        transform: rotate(-160deg) translateY(2px);
+    }
+    55% {
+        transform: rotate(-160deg) translateY(-2px);
+    }
+    70% {
+        transform: rotate(-160deg) translateY(0);
+    }
+    80% {
+        transform: rotate(0);
+    }
+}
